Fix and improve Symbol Export to be Dolphin-readable
It seems the format Dolphin reads is slightly different what the
previous Symbol Export outputted, requiring those section headers.
Additionally, they are split by function and data symbols, so also do it
here.

For convenience, the plugin tries to search the Maps folder of an
Dolphin Emulator user data folder, so it can be quickly loaded into
Dolphin for quick debugging or something.
Also, it remembers the path for an session.